DBZ-1859 Moved metrics to subpackage
